add test file 'rt_file/test/*.rt'

+++ srcs/raytrace/calc_rayvec.c
@@ -14,7 +14,7 @@
 
 double	calc_eye2scrcen_distance(t_camera_info *camera_info)
 {
-	return (WINDOW_WIDTH / (2.0 * tanf(camera_info->fov_deg / 2 * M_PI / 180.0)));
+	return (WINDOW_WIDTH / (2.0 * tan(camera_info->fov_deg / 2 * M_PI / 180.0)));
 }
 
 void	ready_scr2world_ex(t_vec *scr2world_ex, t_camera_info *camera_info)
